Back off dependency-wait defers in sf-queues.
A cluster operation whose dependency was not yet complete was
re-enqueued a flat 15 seconds into the future by the sf-queues
dispatcher, because workitem.py called op.defer() without a delay
and took BaseClusterOperation.defer()'s default. Measured on sfcbr,
essentially all of node_inst_netdesc_op's 15.78s p50 wait on the
user_waiting queue was this poll interval: excluding deferred
operations collapsed it to 0.77s (issue 3863).

Both defer sites in the sf-queues dispatcher (depends_on and
runs_after) now pass a delay drawn from the exponential back-off
schedule sf-net already uses: 0.1s doubling to a 15s cap. Unlike
sf-net, which tracks back-off depth in a per-worker map, sf-queues
jobs are one-shot threads -- so the depth is derived statelessly
from the defer_count the work item already persists across defers.

defer_with_backoff()'s (15, 30, 60) transient-failure schedule is
deliberately untouched; defer()'s delay annotation is widened to
float to match what both daemons now pass. The developer guide's
back-off section gains a note on the sf-queues variant.

Plan alignment: docs/plans/PLAN-cluster-op-visibility.md -- its
future-work list flags this exact user_waiting deferral cadence as
worth a separate fix, which this is.

Fixes #3863

Triggered-By: issue-fix workflow, https://github.com/shakenfist/shakenfist/actions/runs/32996619943

+++ shakenfist/operations/baseoperation.py
@@ -389,7 +389,7 @@ class BaseClusterOperation(BaseOperation):
     def defer(
             self,
             waiting_on: Optional[list[dbo]] = None,
-            delay: int = 15
+            delay: float = 15.0
     ) -> None:
         if not self.queue_name:
             raise CannotDeferUnqueued(
